Fix: MiniProgram Mall and UniformMessage Module
