--- src/MiniProgram/Mall/Client.php
@@ -1,20 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: XueSi
- * Email: <1592328848@qq.com>
- * Date: 2021/4/7
- * Time: 20:52
- */
-
-namespace EasySwoole\WeChat\MiniProgram\Mall;
-
-
-use EasySwoole\WeChat\MiniProgram\BaseClient;
-
-class Client extends BaseClient
-{
-
-
-
-}
